pyproject.toml
[build-system]
# Specifies build tool requirements (standard for setuptools)

# Build system configuration (required by PEP 517/518)
# Tells tools like pip and uv how to build your project if it needs building
# (even simple projects need this section)
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
# PEP 621 metadata
name = "hf-gaia-agent"
version = "0.1.0"
description = "A LangGraph agent using Gemini and various tools to tackle the GAIA benchmark."
authors = [
    { name = "Ahsanul Habib", email = "a.habib.cuet@gmail.com" }, # Replace with your details
]
readme = "README.md"
requires-python = ">=3.9" # Minimum Python version
license = { text = "MIT" } # Or choose another license like Apache-2.0
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License", 
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Runtime dependencies - These are needed for the agent and app to run
dependencies = [
    "beautifulsoup4>=4.9.0",
    "duckduckgo-search>=4.0.0",
    "google-generativeai>=0.4.0",
    "gradio>=4.0.0",
    "groq>=0.4.0",
    "huggingface-hub>=0.20.0",
    "huggingface_hub>=0.14.1",
    "ipykernel>=6.0.0",
    "ipython>=7.0.0",
    "langchain-community>=0.0.20",
    "langchain-core>0.3.0",
    "langchain-experimental",
    "langchain-google-genai>=1.0.0",
    "langchain-groq>=0.1.0",
    "langchain-huggingface>0.1.1",
    "langchain-openai<=0.3.14",
    "langchain>=0.1.0",
    "langchainhub>0.1.15",
    "langgraph>=0.0.30",
    "lark-parser>0.10.0",
    "lorem-text>2.0.0",
    "numpy>=1.21.0",
    "openai>=0.27.0",
    "openpyxl-stubs<=0.1.25",
    "openpyxl>=3.0.0",
    "pandas-stubs>=1.2.0",
    "pandas>=1.5.0",
    "Pillow>=9.0.0",
    "pip>=24.0.0",
    "pytesseract>=0.3.8",
    "python-dotenv>1.0.0",
    "PyYAML-include>=0.1.0",
    "PyYAML>=5.4.1",
    "requests>=2.25.0",
    "streamlit>=1.20.0",
    "tavily-python>=0.3.0",
    "types-PyYAML>=5.4.1",
    "wikipedia>=1.4.0",
    "yt-dlp>=2023.0.0"
]

# Optional dependencies (e.g., for development, testing)
# Install these using 'uv pip install .[dev]'
[project.optional-dependencies]
dev = [
    "ruff",               # Linter/formatter
    "black>=23.0",        # Code formatter
    "isort>=5.0",         # Import sorter
    "mypy>=0.910",        # Type checker
    "pytest>=7.0.0",      # Uncomment if you plan to add pytest tests
    "pytest-cov",         # Uncomment for test coverage
]

# Optional: Project URLs (useful if you publish to PyPI or share widely)
[project.urls]
